add around_about($epoch, $epsilon); rewrote deep_equal
Previously deep_equal was just "data dent these two strings and
compare them to one another". This didn't allow any interesting
hookability on the expected half of things. Now there's a hook
in deep_equal(), which around_about() uses.

This test now works:

deep_ok(
 #  this got is within 90 seconds of Thu Jan 10 09:00:02 2013
  ['Thu Jan 10 09:01:32 2013'], [around_about(1357826402, 90)],
  'close enough'
);

Todo:

When bailing out because it's "not worth checking the full
data structure", we should be able to mutate the $exp'ected
data structure so a failure like this:

deep_ok(
  {hello => 'world', date => $some_date_string},
  {howdy => 'pard, date => around_about($some_epoch, $epsilon)},
);

correctly illustrates a mismatch between the 'hello' and 'howdy'
keys, and doesn't introduce chatter about the 'date' values.